Allow using local references
This allows providing bibliography entries through other ways than cite2url.
This wasn't possible previously, because cite2url would throw an error if a
citation without an url was found. Also it completely overwrote the references
field so we couldn't use that. The former was addressed by providing an option
url2cite-allow-dangling-citations to disable the error and the latter by
concatenating with existing references instead of overwriting them.